# `create_block_generator` refactor