--- a/benchmarks/mempool.py
+++ b/benchmarks/mempool.py
@@ -228,17 +228,17 @@ async def run_mempool_benchmark() -> None:
         print(f"  time: {stop - start:0.4f}s")
         print(f"  per call: {(stop - start) / total_bundles * 1000:0.2f}ms")
 
-        print("\nProfiling create_bundle_from_mempool()")
+        print("\nProfiling create_block_generator()")
         with enable_profiler(True, f"create-{suffix}"):
             start = monotonic()
-            for _ in range(500):
-                await mempool.create_bundle_from_mempool(
+            for _ in range(50):
+                await mempool.create_block_generator(
                     last_tb_header_hash=rec.header_hash,
                     get_unspent_lineage_info_for_puzzle_hash=get_unspent_lineage_info_for_puzzle_hash,
                 )
             stop = monotonic()
         print(f"  time: {stop - start:0.4f}s")
-        print(f"  per call: {(stop - start) / 500 * 1000:0.2f}ms")
+        print(f"  per call: {(stop - start) / 50 * 1000:0.2f}ms")
 
         print("\nProfiling new_peak() (optimized)")
         blocks: list[tuple[BenchBlockRecord, list[bytes32]]] = []