--- a/chia/full_node/mempool.py
+++ b/chia/full_node/mempool.py
@@ -473,10 +473,10 @@ class Mempool:
 
     async def create_bundle_from_mempool_items(
         self,
-        item_inclusion_filter: Callable[[bytes32], bool],
         get_unspent_lineage_info_for_puzzle_hash: Callable[[bytes32], Awaitable[Optional[UnspentLineageInfo]]],
         constants: ConsensusConstants,
         height: uint32,
+        item_inclusion_filter: Optional[Callable[[bytes32], bool]] = None,
     ) -> Optional[tuple[SpendBundle, list[Coin]]]:
         cost_sum = 0  # Checks that total cost does not exceed block maximum
         fee_sum = 0  # Checks that total fees don't exceed 64 bits
@@ -506,7 +506,7 @@ class Mempool:
             if current_time - bundle_creation_start > 1:
                 log.info(f"exiting early, already spent {current_time - bundle_creation_start:0.2f} s")
                 break
-            if not item_inclusion_filter(name):
+            if item_inclusion_filter is not None and not item_inclusion_filter(name):
                 continue
             try:
                 assert item.conds is not None