This PR is best reviewed one commit at a time.

### Purpose:

Change the mempool interface to create a block generator, rather than a large spend bundle (to then be turned into a block generator)

The main reason for this is to prepare for changing the implementation of the block generator creation to use an incrementally compressed block creator, to fill block more effectively.

It has some other benefits I think:

1. The spend bundle is an unnecessary step, skipping it should save some time and space
2. It makes it easier to unit test the block generator creation (which currently requires interacting with `full_node_api`, and probably requires a full simulation environment.

The first commit is only tangentially related, and simplifies the function interface related to mempool item inclusion.

### Current Behavior:

The mempool has a method: `create_bundle_from_mempool_items()`

### New Behavior:

The mempool has a method: `create_block_generator`

### benchmarks

This is not expected to affect performance, but the new function combines picking transactions and serializing the block, so it will take longer than just picking the transactions. The benchmarks are now:

Serializing the block with compression is especially expensive:

before:
```
Profiling create_bundle_from_mempool()
  output written to: mempool-create-mt.png
  time: 4.6944s
  per call: 9.39ms
```

after:
```
Profiling create_block_generator()
  output written to: mempool-create-mt.png
  time: 43.8237s
  per call: 876.47ms
```